#66: 세션 멈춤을 해결했다.
