+++ MC2-TEAM10-BURIBURI/Camera/TutorialCameraView.swift
@@ -8,7 +8,7 @@ struct TutorialCameraView: View {
     @AppStorage("isFirstLaunch") private var isFirstLaunch = true
     @StateObject private var photomodel = PhotoDataModel()
     @EnvironmentObject var dataModel: DataModel
-    @EnvironmentObject var heavyViewStatusModel: HeavyViewStatusModel
+    @EnvironmentObject var arViewStatusModel: ARViewStatusModel
     
     @State var showsDecisionPage: Bool = false // true -> 선택하는 화면 띄움
     @State var returnedTuple: (Data, [CGPoint]) = (Data(), [CGPoint]()) // convertToBackgroundRemovedPNGDataAndPointArray 함수의 반환값
@@ -22,60 +22,57 @@ struct TutorialCameraView: View {
     var body: some View {
         NavigationStack {
             GeometryReader { geometry in
-                if heavyViewStatusModel.tutorialCameraViewIsDrawn {
-                    ZStack {
-                        if !showsDecisionPage {
-                            ViewfinderView(image:  $photomodel.viewfinderImage )
-                                .overlay(alignment: .top) {
+                ZStack {
+                    if !showsDecisionPage {
+                        ViewfinderView(image:  $photomodel.viewfinderImage )
+                            .overlay(alignment: .top) {
+                                Color.black
+                                    .opacity(0)
+                                    .frame(height: geometry.size.height * Self.barHeightFactor)
+                            }
+                            .overlay(alignment: .topLeading) {
+                                buttonsView2()
+                                    .frame(height: geometry.size.height * Self.barHeightFactor)
+                                    .background(.black.opacity(0))
+                            }
+                            .overlay(alignment: .bottom) {
+                                buttonsView()
+                                    .frame(height: geometry.size.height * Self.barHeightFactor)
+                                    .background(.black.opacity(0.2))
+                            }
+                        
+                            .overlay(alignment: .center)  {
+                                Color.clear
+                                    .frame(height: geometry.size.height * (1 - (Self.barHeightFactor * 2)))
+                                    .accessibilityElement()
+                                    .accessibilityLabel("View Finder")
+                                    .accessibilityAddTraits([.isImage])
+                            }
+                            .background(.black)
+                    } else {
+                        ZStack {
+                            ViewfinderView(image: $photomodel.viewfinderImage)
+                                .overlay {
                                     Color.black
-                                        .opacity(0)
-                                        .frame(height: geometry.size.height * Self.barHeightFactor)
+                                        .opacity(0.7)
                                 }
-                                .overlay(alignment: .topLeading) {
-                                    buttonsView2()
-                                        .frame(height: geometry.size.height * Self.barHeightFactor)
-                                        .background(.black.opacity(0))
+                                .overlay(alignment: .center) {
+                                    PNGImageDataView(pngData: returnedTuple.0)
+                                        .frame(width: getWidth() * 0.5)
                                 }
                                 .overlay(alignment: .bottom) {
-                                    buttonsView()
+                                    buttonsView3()
                                         .frame(height: geometry.size.height * Self.barHeightFactor)
-                                        .background(.black.opacity(0.2))
+                                }
+                                .overlay(alignment: .topTrailing) {
+                                    buttonsView4()
+                                        .frame(height: geometry.size.height * Self.barHeightFactor)
                                 }
                             
-                                .overlay(alignment: .center)  {
-                                    Color.clear
-                                        .frame(height: geometry.size.height * (1 - (Self.barHeightFactor * 2)))
-                                        .accessibilityElement()
-                                        .accessibilityLabel("View Finder")
-                                        .accessibilityAddTraits([.isImage])
-                                }
-                                .background(.black)
-                        } else {
-                            ZStack {
-                                ViewfinderView(image: $photomodel.viewfinderImage)
-                                    .overlay {
-                                        Color.black
-                                            .opacity(0.7)
-                                    }
-                                    .overlay(alignment: .center) {
-                                        PNGImageDataView(pngData: returnedTuple.0)
-                                            .frame(width: getWidth() * 0.5)
-                                    }
-                                    .overlay(alignment: .bottom) {
-                                        buttonsView3()
-                                            .frame(height: geometry.size.height * Self.barHeightFactor)
-                                    }
-                                    .overlay(alignment: .topTrailing) {
-                                        buttonsView4()
-                                            .frame(height: geometry.size.height * Self.barHeightFactor)
-                                    }
-                                
-                            }
                         }
                     }
-                } else {
-                    EmptyView()
                 }
+                
             }
             .task {
                 await photomodel.camera.start()
@@ -87,12 +84,6 @@ struct TutorialCameraView: View {
             .navigationBarHidden(true)
             .ignoresSafeArea()
             .statusBar(hidden: true)
-            .onAppear {
-                heavyViewStatusModel.resetTutorialCameraView()
-            }
-            .onDisappear {
-                heavyViewStatusModel.hideTutorialCameraView()
-            }
         }
         .navigationBarHidden(true)
     }
@@ -173,7 +164,7 @@ struct TutorialCameraView: View {
         HStack {
             Spacer()
                 .frame(width: getWidth() * 0.05)
-            NavigationLink(destination: ARView().environmentObject(heavyViewStatusModel)) {
+            NavigationLink(destination: ARView().environmentObject(arViewStatusModel)) {
                 HStack {
                     Image(systemName: "chevron.left")
                         .foregroundColor(Color.white)
