+++ MC2-TEAM10-BURIBURI/Model/HeavyViewStatusModel.swift
@@ -1,0 +1,39 @@
+//
+//  ARViewStatusModel.swift
+//  MC2-TEAM10-BURIBURI
+//
+//  Created by Wonil Lee on 12/21/23.
+//
+
+import Foundation
+
+final class ARViewStatusModel: ObservableObject {
+    @Published var arViewIsDrawn = true
+
+    @MainActor
+    func setARViewIsDrawnTrue() {
+        arViewIsDrawn = true
+    }
+    @MainActor
+    func setARViewIsDrawnFalse() {
+        arViewIsDrawn = false
+    }
+    
+    func resetARView() {
+        Task(priority: .userInitiated) {
+            await setARViewIsDrawnFalse()
+            do {
+                try await Task.sleep(nanoseconds: 10_000_000)
+            } catch {
+                print("error sleeping: \(error.localizedDescription)")
+            }
+            await setARViewIsDrawnTrue()
+        }
+    }
+
+    func hideARView() {
+        Task {
+            await setARViewIsDrawnFalse()
+        }
+    }
+}
