+++ MC2-TEAM10-BURIBURI/View/StartView2.swift
@@ -12,7 +12,7 @@ import UIKit
 struct StartView2: View {
     
     @EnvironmentObject var dataModel : DataModel
-    @EnvironmentObject private var heavyViewStatusModel: HeavyViewStatusModel
+    @EnvironmentObject private var arViewStatusModel: ARViewStatusModel
     
     var body: some View {
         SummonGIF("Title_00") //화면비 수정해야 함
@@ -22,7 +22,7 @@ struct StartView2: View {
                 DispatchQueue.main.asyncAfter(deadline: .now() + 5) {
                     
                     //애니메이션이 있어야 할 듯..
-                    UIApplication.shared.windows.first?.rootViewController = UIHostingController(rootView: ARView().environmentObject(dataModel).environmentObject(heavyViewStatusModel))
+                    UIApplication.shared.windows.first?.rootViewController = UIHostingController(rootView: ARView().environmentObject(dataModel).environmentObject(arViewStatusModel))
                 }
             }
     }
